pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-cli"
version = "0.1.0"
description = "CLI for managing AI agents, rules, and skills from a centralized registry"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "typer[all]>=0.12.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.scripts]
forge = "forge.cli.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
